requirements.txt
# KYC Client Onboarding Intelligence System - Requirements
# Python 3.10+ required

# Core dependencies
anthropic>=0.40.0,<1.0.0      # Anthropic API client
httpx>=0.27.0,<1.0.0          # Async HTTP client
pydantic>=2.0.0,<3.0.0        # Data validation

# Output generation
rich>=13.0.0,<14.0.0          # Rich terminal output
fpdf2>=2.7.0,<3.0.0           # PDF generation

# HTML parsing (optional but recommended)
beautifulsoup4>=4.12.0,<5.0.0 # HTML parsing
lxml>=5.0.0,<6.0.0            # Fast HTML/XML parser (optional)

# Sanctions list fuzzy matching
rapidfuzz>=3.0.0,<4.0.0       # Fuzzy string matching for screening

# Environment configuration (optional)
python-dotenv>=1.0.0,<2.0.0   # .env file loading

# Development dependencies (optional)
# pytest>=8.0.0,<9.0.0        # Testing framework
# pytest-asyncio>=0.23.0      # Async test support
# pytest-cov>=4.0.0           # Coverage reporting
